Add files via upload
This document represents the result of the research and practice of database establishment for the companies in the retail industry, it also includes all work associated with preparation, design, and implementation of the database. This document and the corresponding data files and appendix are intended to enhance the understanding of the entire project. To meet the general business requirement for the companies in the retail industry, the system requirements will indicate the required hardware and software including but not limited to CPU, MySQL, and Ubuntu. Major entities within the system and the inner relationship amount will be shown by the visualization, database structure, and data dictionary. Moreover, a security backup plan is also included in this project as part of the support for recovering the database system if necessary.
 
The purpose and goal for this project are to illustrate a detailed plan for implement a database for the retail companies, which by building a searchable database, entity-relationship diagram (ERD), operational rules, and related elements to manage various resources and meet the market trend for business development.